+++ lib/data/models/restaurants_model/data.dart
@@ -49,8 +49,8 @@ class Data {
   }
   int? id;
   String? address;
-  int? coordinateY;
-  int? coordinateX;
+  double? coordinateY;
+  double? coordinateX;
   String? cityName;
   String? nationName;
   String? typeOfFood;
