+++ lib/data/models/search_restaurants_model/resturants.dart
@@ -33,8 +33,8 @@ class Resturants {
   }
   int? id;
   String? address;
-  int? coordinateY;
-  int? coordinateX;
+  double? coordinateY;
+  double? coordinateX;
   String? cityName;
   String? nationName;
   String? typeOfFood;
