+++ lib/data/models/search_tourist_dis_model/attraction_activities.dart
@@ -27,8 +27,8 @@ class AttractionActivities {
   }
   int? id;
   String? address;
-  int? coordinateY;
-  int? coordinateX;
+  double? coordinateY;
+  double? coordinateX;
   String? cityName;
   String? nationName;
   String? attractionActivityName;
